+++ 2key-protocol/src/campaignValidator/interfaces.ts
@@ -1,5 +1,5 @@
 export interface ITwoKeyCampaignValidator {
     validateCampaign: (campaignAddress: string, nonSingletonHash: string, from:string) => Promise<string>,
     isCampaignValidated: (campaignAddress:string) => Promise<boolean>,
-    getCampaignNonSingletonHash: (campaignAddress:string) => Promise<string>,
+    getCampaignNonSingletonsHash: (campaignAddress:string) => Promise<string>,
 }